requirements.txt
streamlit==1.24.0
opencv-python
numpy==1.24.3
easyocr==1.7.0
pillow==9.5.0
scikit-image==0.21.0
scipy==1.11.1